feat:notification_user
user cannot delete himself and is notified of this

Kaique A Mesquita

+++ frontend/Estoque_WebApp/src/app/pages/user/users-view/users-view.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
+import { AfterViewInit, Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
 import { PTableComponent } from '../../../shared/components/p-table/p-table.component';
 import { UserService } from './../../../core/services/user.service';
 import { User } from './../../../shared/models/user';
@@ -7,17 +7,22 @@ import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { AuthenticationService } from './../../../core/authentication/authentication.service';
 import { NavigationWatcherService } from '../../../core/services/navigation-watcher.service';
+import { ModalModule } from '../../../shared/modules/modal/modal.module';
+import { ModalComponent } from '../../../shared/components/modal/modal.component';
+
 
 @Component({
   selector: 'app-users-view',
-  imports: [PTableComponent, CommonModule],
+  imports: [PTableComponent, CommonModule, ModalModule],
  standalone: true,
 templateUrl: './users-view.component.html',
   styleUrl: './users-view.component.css'
 })
-export class UsersViewComponent implements OnInit, OnDestroy {
+export class UsersViewComponent implements OnInit, OnDestroy, AfterViewInit {
   users: User[] = []; // Array para armazenar os usuários
-    private navigationSub?: Subscription;
+  private navigationSub?: Subscription;
+  isDeleted: boolean = false; // Variável para controlar a exibição do modal de exclusão
+  @ViewChild('exceptionUser') exceptionUser!: ModalComponent;
 
 
   constructor(private userService :UserService, private auth: AuthenticationService, 
@@ -41,6 +46,9 @@ export class UsersViewComponent implements OnInit, OnDestroy {
     // Ouve eventos de navegação e recarrega se a rota atual for a deste componente
   
   }
+
+    ngAfterViewInit(): void {}
+
   /**
     Método para obter o token de autenticação do usuário.
    retorna O token de autenticação ou null se não estiver autenticado.
@@ -68,14 +76,32 @@ export class UsersViewComponent implements OnInit, OnDestroy {
     });
   }
 
+
   // Método para deletar um usuário
   DeleteUser(userId: number): void {
-   try {
-    this.userService.deleteUser(userId); // Chama o serviço para deletar o usuário
-    this.users = this.users.filter(user => user.id !== userId); // Atualiza a lista de usuários removendo o usuário deletado
-   }catch (error) {
-    console.error('Erro ao deletar usuário:', error);
-   }
+  const token = this.auth.decodeToken();
+  if (token && token.sub) {
+    //pegar o nome do usuário na lista  user pelo id
+    const userName = this.users.find(user => user.id === userId)?.name;
+
+    console.log('Nome do usuário:', userName);
+    console.log('Usuário logado:', token.sub);
+    if (userName === token.sub) {
+      // Se o usuário for o mesmo que está logado, não permite a exclusão
+      console.error('Não é possível deletar o próprio usuário.');
+      this.exceptionUser.toggle();
+
+      return;
+    }else{
+        try {
+            this.userService.deleteUser(userId); // Chama o serviço para deletar o usuário
+            this.users = this.users.filter(user => user.id !== userId); // Atualiza a lista de usuários removendo o usuário deletado
+          }catch (error) {
+            console.error('Erro ao deletar usuário:', error);
+          }
+    }
+  }
+   
   }
 
   // Método para redirecionar para a página de edição de usuário
